Refactored backend to separate graphbrain-specific behavior. work-in-progress, API and some functions still not updated. 	renamed:    api/parser.py -> api/graphbrain_parser.py         modified:   newpotato/datatypes.py 	new file:   newpotato/extractors/extractor.py 	new file:   newpotato/extractors/graphbrain_extractor.py 	renamed:    newpotato/parser.py -> newpotato/extractors/graphbrain_parser.py 	modified:   newpotato/hitl.py 	modified:   newpotato/term_client.py 	modified:   newpotato/utils.py
